Change ProcStat vsize/rss from int32 to int64 to prevent overflow
Agent-Logs-Url: https://github.com/open-gpdb/yagpcc/sessions/6a34e218-f71b-462f-9a54-f59a5700fcdb

+++ internal/utils/procfs.go
@@ -127,7 +127,7 @@ func convertProcStat(s *procfs.ProcStat) *pb.ProcStat {
 		NumThreads:          int32(s.NumThreads),
 		Starttime:           int64(s.Starttime),
 		Vsize:               int64(s.VSize),
-		Rss:                 int32(s.RSS),
+		Rss:                 int64(s.RSS),
 		RssLimit:            int64(s.RSSLimit),
 		Processor:           int32(s.Processor),
 		RtPriority:          int32(s.RTPriority),
